requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2.5
django-background-tasks==1.2.8
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
rest-framework-simplejwt==0.0.2
setuptools==80.3.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
